+++ src/interface.d.ts
@@ -4,6 +4,11 @@ export interface ElectronAPI {
   on(channel: string, func: (...args: any[]) => void): void
   invoke(channel: string, data?: unknown): Promise<any>
   getFilePath(file: File): string
+  store: {
+    get(key: string, defaultValue?: any): Promise<any>
+    set(key: string, value: any): Promise<boolean>
+    remove(key: string): Promise<boolean>
+  }
 }
 
 declare global {