+++ src/lib/components/layout/Sidebar.svelte
@@ -1,36 +1,25 @@
 <script lang="ts">
   import {
+    addNetworkNode,
     getAvailableNodes,
     getStoredNetworkNodes,
-    setNetworkNodes,
-    setRegistry,
+    removeNetworkNode,
   } from '../../stores/nodes.svelte'
   import { dndNodeDataState } from '../../stores/dndStore.svelte'
-  import defaultNodesJson from '../../data/defaultNodes.json'
-  import defaultNetworkNodesJson from '../../data/defaultNetworkNodes.json'
   import {
     nodeColors,
     NodeType,
     returnNodeName,
     type NodeData,
-    type RegisteredNodes,
   } from '../../types/nodeTypes'
   import { fade } from 'svelte/transition'
   import { sideBarState } from '../../stores/sidebar.svelte'
-
-  const defaultNodes = defaultNodesJson as RegisteredNodes
-  const defaultNetworkNodes = defaultNetworkNodesJson as RegisteredNodes
+  import { toastState } from '../../stores/toastsStore.svelte'
 
   let isMouseOver = $state(false)
   const showNodeNames = $derived(isMouseOver || sideBarState.isExpanded)
 
-  if (defaultNodes) {
-    setRegistry(defaultNodes)
-  }
   const availableNodes = $derived(getAvailableNodes())
-  if (defaultNetworkNodes) {
-    setNetworkNodes(defaultNetworkNodes)
-  }
   const storedNetworkNodes = $derived(getStoredNetworkNodes())
 
   const onDragStart = (event: DragEvent, node: NodeData) => {
@@ -43,6 +32,18 @@
 
   const returnNodeColor = (nodeTypeName) => {
     return nodeColors[nodeTypeName]
+  }
+
+  const handleDelete = async (networkNodeName) => {
+    try {
+      await removeNetworkNode(networkNodeName)
+    } catch (e) {
+      toastState.add({
+        message: e.message || `Failed to delete node ${addNetworkNode}`,
+        type: 'error',
+      })
+      console.error(`Failed to delete node ${addNetworkNode}`, e.message)
+    }
   }
 </script>
 
@@ -84,6 +85,16 @@
             <span transition:fade|global={{ duration: 250 }}>
               {returnNodeName(node)}
             </span>
+            <!-- svelte-ignore a11y_click_events_have_key_events -->
+            <svg
+              class="close"
+              onclick={() => handleDelete(node.name)}
+              viewBox="0 0 12 12"
+            >
+              <circle cx="6" cy="6" r="6" />
+              <line x1="3" y1="3" x2="9" y2="9" />
+              <line x1="9" y1="3" x2="3" y2="9" />
+            </svg>
           {/if}
         </div>
       {/each}
@@ -116,6 +127,7 @@
     transition: height 0.25s ease;
   }
   .node {
+    position: relative;
     padding: 0.5rem 1rem;
     margin: 0 1rem;
     border-radius: 5px;
@@ -125,5 +137,25 @@
 
   .node:hover {
     border-color: var(--border-color-hover);
+  }
+
+  .close {
+    position: absolute;
+    top: -10px;
+    right: -10px;
+    width: 20px;
+    height: 20px;
+    cursor: pointer;
+    fill: #f44;
+    transition: transform 0.3s ease;
+  }
+
+  .close:hover {
+    transform: scale(1.5);
+  }
+
+  .close line {
+    stroke: #fff;
+    stroke-width: 2;
   }
 </style>